improve locator plugin api
