+++ packages/locator/src/hoverElement.ts
@@ -4,13 +4,13 @@ import { makeEventListener } from "@solid-primitives/event-listener"
 
 // TODO: better support touch
 
-export function makeHoverElementListener(onHover: (el: Element | null) => void): void {
-  let last: Element | null = null
+export function makeHoverElementListener(onHover: (el: HTMLElement | null) => void): void {
+  let last: HTMLElement | null = null
   const handleHover = (e: { target: unknown }) => {
     const { target } = e
-    if (target === last || (!(target instanceof Element) && target !== null)) return
+    if (target === last || (!(target instanceof HTMLElement) && target !== null)) return
     onHover((last = target))
   }
-  makeEventListener(window, "pointerover", handleHover)
-  makeEventListener(document, "pointerleave", handleHover.bind(void 0, { target: null }))
+  makeEventListener(window, "mouseover", handleHover)
+  makeEventListener(document, "mouseleave", handleHover.bind(void 0, { target: null }))
 }
