+++ packages/locator/tsup.config.ts
@@ -1,3 +1,3 @@
 import defineConfig from "../../configs/tsup.config"
 
-export default defineConfig({ extension: "tsx", server: true })
+export default defineConfig({ server: true })
