+++ playgrounds/sandbox/src/index.tsx
@@ -1,5 +1,6 @@
 /* @refresh reload */
 import { render } from "solid-js/web"
+import "solid-devtools"
 import { useLocator } from "solid-devtools"
 
 import App from "./App"
